Refactored licenses class.
- Merged load() and get() functions.
- ids() function no longer requires get() to preload the dictionary
- Fixed example scripts
- Changed SHOW_ALL macro declaration from int to uint16_t

+++ pyslurm/slurm.pxd
@@ -392,7 +392,7 @@ cdef extern from 'slurm/slurm.h' nogil:
         NODE_STATE_FUTURE
         NODE_STATE_END
 
-    int SHOW_ALL
+    uint16_t SHOW_ALL
 
     ctypedef enum acct_energy_type:
         ENERGY_DATA_JOULES_TASK
